html/interaction/focus/focusgroup/tentative/backward-navigation/moves-to-previous-item-out-of-shadow-dom.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: focusgroup - Focus moves to previous item out of Shadow DOM.</title>
<link rel="author" title="Microsoft" href="http://www.microsoft.com/">
<link rel="help" href="https://github.com/MicrosoftEdge/MSEdgeExplainers/blob/main/Focusgroup/explainer.md">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="../resources/focusgroup-utils.js"></script>

<div id=root focusgroup>
  <span id=item1 tabindex=0>item1</span>
  <div id=fg2 focusgroup=extend>
    <template shadowroot=open>
      <span id=item2 tabindex=-1>item2</span>
    </template>
  </div>
</div>

<script>

  promise_test(async t => {
    var item1 = document.getElementById("fg2").shadowRoot.getElementById("item1");
    var item2 = document.getElementById("item2");

    await focusAndKeyPress(item2, kArrowUp);
    assert_equals(document.activeElement, item1);

    await focusAndKeyPress(item2, kArrowLeft);
    assert_equals(document.activeElement, item1);
  }, "When the focus is set on a focusgroup item, an arrow key press should move the focus to the previous item even if the item we were on was in a ShadowDOM and the previous item is not.");

</script>
